backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.0.0
icecream==2.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.2.2
Pygments==2.18.0
PyJWT==2.8.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.2
Werkzeug==3.0.3
